+++ apps/dbagent/src/evals/chat/toolChoice.eval.ts
@@ -1,8 +1,8 @@
 import { afterAll, beforeAll, describe, expect, vi } from 'vitest';
-import { evalChat } from '~/evals/lib/chatRunner';
+import { evalChat } from '~/evals/lib/chat-runner';
 import * as connectionInfoExports from '~/lib/db/connection-info';
 import * as projectsExports from '~/lib/db/projects';
-import { PostgresConfig, runSql, startPostgresContainer } from '../lib/evalDockerDb';
+import { PostgresConfig, runSql, startPostgresContainer } from '../lib/eval-docker-db';
 import { EvalCase, runEvals } from '../lib/vitestHelpers';
 
 vi.spyOn(projectsExports, 'getProjectById').mockImplementation(async (id) => {
